requirements/base.txt
chargebee>=2,<3
coreapi==2.3.3
Django==1.11.3
django-crispy-forms==1.6.1
django-filter==1.1.0
django-guardian==1.4.9
django-import-export==0.5.1
git+https://github.com/Humanitec/django-oauth-toolkit-jwt@v0.4.1#egg=django-oauth-toolkit-jwt
djangorestframework==3.7.3
psycopg2-binary>=2.7,<2.8
pytz==2018.3
git+https://github.com/toladata/social-core#egg=social-core
social-auth-app-django==2.1.0
git+https://github.com/toladata/TolaActivity@v2.15.8#egg=tola-activity
voluptuous==0.10.5
django-health-check==3.5.1
futures==3.2.0
graphene-django==2.1.0
django-cors-headers==2.1.0
